little more x64 fixes

+++ codemp/cgame/cg_main.c
@@ -237,7 +237,7 @@ intptr_t vmMain( int command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intpt
 											  //garbage. In other words, use at your own risk.
 
 	case CG_GET_MODEL_LIST:
-		return (int)cgs.gameModels;
+		return /*(int)*/cgs.gameModels;
 
 	case CG_CALC_LERP_POSITIONS:
 		CG_CalcEntityLerpPositions( &cg_entities[arg0] );
